Improve test coverage.
This patch exercises all the branches of the alt
operator. We also update the run-kcov script so that
it will work for a clean checkout of the project.
Previously it assumed that someone had already run
`cargo test` to create a binary. Now `cargo clean && ./ci/run-kcov`
should work.

When running locally, you can see the coverage report by going
to `target/cov/index.html`.