Add a handrolled lexer
LALRPOP will generate a lexer for you by examining your inline
terminal regex, but there is no obvious way to implement certain
features such as comments or escape sequences in literals[1]. In order
to get those features, we have to roll our own lexer. This patch
adds a handrolled lexer based roughly on the one found in the
Gluon crate.

[1]: If LALRPOP used the regex crate we could do escape sequences with
a little bit of post-processing, but for some reason LALRPOP has its own
regex implementation (using the same syntax) and does not support a
number of features such as lazy matching.