Format with rustfmt
This patch establishes the style standards for the remake
project. Mostly we are going to use defaults. The biggest
departure from defaults is setting the max line width to
80 because Linus Was Right. The real reason is just that I
like to have my code on one half of the screen and my
`watchexec -e rs,lalrpop "cargo test"` command on the other
half, and the default was causing way to much line wrapping.
If you have found this commit message through `git blame`
because the low column count is forcing your code to take
up a lot of vertical space, we can talk about changing it.

Style policy does not exist unless it is in CI, so this
patch also adds a check to travis to make sure that
things follow the "right" style.