Add support for integer and floating literals
This patch adds support for integer and floating point literals.
We only support 64-bit versions of each for the sake of simplicity.
The typestrings presented to users will just be "int" and "string".

This patch includes changes across the whole language stack, from
lexer to evaluator.

These won't start to be useful until support for more types and
some actual control flow lands, but its a good start.