Switch interpreter to use reference counting
Up till now, none of the features we have added have required
garbage collection of any sort. That's about to change when we
add the ability to use structured data types like tuples, lists
and hashtables. This patch is preparing for that.

This patch also adds an error for division by zero, which was
an oversight of the previous implimentation.