Add dict and tuple types
This patch adds initial support for dicts and tuple
types. The patch only adds basic language-level operators
to manipulate those types, but does not include methods
for working with them (such as a .keys() and .items()) methods.

Despite some comments that I made in #15, I decided to add tuples
because they are importaint for representing returned error values
in a way that is reasonable. Remake will never get an exception
mechanism, so tuple return values are the best choice for error
values. An unrecoverable error on a failed dict access is obviously
the wrong thing to do.

Closes #15.
Closes #16.