Add support for if statements and expressions
This patch adds support for if expressions, which must
always have both a true and a false branch and if statements
with may omit the false branch. Both may have any number of
`else if` clauses. Remake if statements differ from rust if
statements in that the conditional must be wrapped in parens.
This is not ideal, but is a requirement for compatibility with
postfix regex operators.