Add builtin system
This patch adds a system for registering builtin functions and
adds several such functions for dealing with some of the builtin
data types though the functions added in this patch are not
meant to be complete.

The work on this patch revealed some issues with the way that
I was using the Rc pointers to various values. I should not
keep any Borrows around across calls to `eval_` or `exec`,
but we currently do.